+++ swing/framework-ui-test/src/main/java/is/codion/swing/framework/ui/test/EntityEditPanelTestUnit.java
@@ -21,6 +21,8 @@ package is.codion.swing.framework.ui.test;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
 
+import java.util.function.Function;
+
 import static java.util.Objects.requireNonNull;
 
 /**
@@ -29,17 +31,17 @@ import static java.util.Objects.requireNonNull;
 public class EntityEditPanelTestUnit {
 
 	private final SwingEntityEditModel editModel;
-	private final Class<? extends EntityEditPanel> editPanelClass;
+	private final Function<SwingEntityEditModel, EntityEditPanel> editPanel;
 
 	/**
-	 * Instantiates a new edit panel test unit for the given edit panel class
+	 * Instantiates a new edit panel test unit for the given edit panel
 	 * @param editModel the edit model
-	 * @param editPanelClass the edit panel class
+	 * @param editPanel provides the edit panel to test
 	 */
 	protected EntityEditPanelTestUnit(SwingEntityEditModel editModel,
-																		Class<? extends EntityEditPanel> editPanelClass) {
+																		Function<SwingEntityEditModel, EntityEditPanel> editPanel) {
 		this.editModel = requireNonNull(editModel, "editModel");
-		this.editPanelClass = requireNonNull(editPanelClass, "editPanelClass");
+		this.editPanel = requireNonNull(editPanel, "editPanel");
 	}
 
 	/**
@@ -47,22 +49,13 @@ public class EntityEditPanelTestUnit {
 	 * @throws Exception in case of an exception
 	 */
 	protected final void testInitialize() throws Exception {
-		createEditPanel().initialize();
-	}
-
-	/**
-	 * Creates the edit panel for testing
-	 * @return the edit panel to test
-	 * @throws Exception in case of an exception
-	 */
-	protected EntityEditPanel createEditPanel() throws Exception {
-		return editPanelClass.getConstructor(SwingEntityEditModel.class).newInstance(editModel());
+		editPanel.apply(editModel).initialize();
 	}
 
 	/**
 	 * @return the edit model
 	 */
-	protected SwingEntityEditModel editModel() {
+	protected final SwingEntityEditModel editModel() {
 		return editModel;
 	}
 }