+++ changelog.md
@@ -14,6 +14,8 @@ Codion Change Log
 - SchemaDomain.SchemaSettings.primaryKeyColumnSuffix() added for improved foreign key names.
 ### is.codion.tools.generator.domain
 - DomainSource, domain implementation now in base source package with the API in a separate api package.
+### is.codion.swing.framework.ui.test
+- EntityEditPanelTestUnit, createEditPanel() removed, edit panel initializer function parameter added.
 
 ## 0.18.9
 ### is.codion